Move build-stack-docs cli to bin subpackage
This lets the build module be CLI agnostic (and also removes the
distraction of the deprecated build-stack-docs CLI).
