+++ setup.py
@@ -84,7 +84,7 @@ setup_requires = [
 console_scripts = [
     'stack-docs = documenteer.stackdocs.stackcli:main',
     'package-docs = documenteer.stackdocs.packagecli:main',
-    'build-stack-docs = documenteer.stackdocs.build:run_build_cli',
+    'build-stack-docs = documenteer.bin.buildstackdocs:run_build_cli',
     'refresh-lsst-bib = documenteer.bin.refreshlsstbib:run'
 ]
